Highlight that the amount to stake is in LPTU (#145)
Community members adrianm and Twodawg both reported in Livepeer Discord that they found the LPTU vs. LPT distinction confusing, and that they unnecessarily burned gas.

This PR seeks to highlight this point, by adding a bold "Important note:" text, so that others are more likely to see it, and avoid wasting gas.