home/hero: Use hero videos hosted in Livepeer Studio
We've had a bandwidth surge in the last few days, and almost
2TB of it was downloading these videos from what I assume is
some Vercel static storage.

I've uploaded each video to Livepeer instead to our shared
team@livepeer.org account (1pwd). Replaced the links of the videos
here with the download URLs of them in the VoD API.

Notice I'm not using HLS nor even our player though, as that would
be a bigger frontend change I'm not able to do right now. Would
welcome anyone to do that follow-up tho as it'd bring some further
improvements and showcase more of our tech in the homepage :)